moved bookmarks helper, and notice publisher
