+++ Vice/app/src/main/java/martell/com/vice/activity/MainActivity.java
@@ -31,10 +31,10 @@ import martell.com.vice.nav_drawer.NavDrawerDivider;
 import martell.com.vice.nav_drawer.NavDrawerEntry;
 import martell.com.vice.nav_drawer.NavDrawerItem;
 import martell.com.vice.nav_drawer.NavDrawerToggle;
-import martell.com.vice.nav_drawer.NotificationPublisher;
+import martell.com.vice.helpers.NotificationPublisher;
 import martell.com.vice.R;
 import martell.com.vice.adapters.ViewPagerAdapter;
-import martell.com.vice.dbHelper.NotificationDBHelper;
+import martell.com.vice.helpers.NotificationDBHelper;
 import martell.com.vice.fragment.LatestNewFragment;
 import martell.com.vice.fragment.NavigationDrawerFragment;
 import martell.com.vice.services.ViceAPIService;
