+++ Vice/app/src/main/java/martell/com/vice/fragment/LatestNewFragment.java
@@ -24,11 +24,10 @@ import java.util.Arrays;
 import jp.wasabeef.recyclerview.adapters.AlphaInAnimationAdapter;
 import martell.com.vice.activity.ArticleActivity;
 import martell.com.vice.adapters.ArticleAdapter;
-import martell.com.vice.BookmarksHelper;
+import martell.com.vice.helpers.BookmarksHelper;
 import martell.com.vice.activity.MainActivity;
 import martell.com.vice.R;
-import martell.com.vice.RV_SpaceDecoration;
-import martell.com.vice.dbHelper.DatabaseHelper;
+import martell.com.vice.helpers.DatabaseHelper;
 import martell.com.vice.models.Article;
 import martell.com.vice.models.ArticleArray;
 import martell.com.vice.services.ViceAPIService;
